Use READ_PHONE_NUMBERS instead of READ_PHONE_STATE when API >= 30
When the app needs to ask permission for the device's phone number, the PermissionsProvider will
now check if the API level is >= 30 and use the appropriate permission: READ_PHONE_NUMBERS if 30+,
READ_PHONE_STATE if lower than 30. This change is transparent to anything outside of
PermissionsProvider.

Also changed the AndroidManifest to reflect the change in necessary permissions.